+++ src/components/common/IconButtons.tsx
@@ -1,4 +1,4 @@
-import { Button } from "./ui/button";
+import { Button } from "../ui/button";
 interface props {
 	src: string;
 	alt: string;
